feat: "Comment 생성, 삭제 회원 검증 로직 추가" #10

+++ src/main/java/com/trip/comment/service/CommentServiceImpl.java
@@ -20,8 +20,18 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void deleteComment(int commentId) {
+    public boolean deleteComment(int commentId, String loggedInEmail) {
+        // 댓글 작성자 이메일 조회
+        String authorEmail = commentMapper.findAuthorEmailByCommentId(commentId);
+
+        // 검증 로직
+        if (!loggedInEmail.equals(authorEmail)) {
+            return false; // 작성자와 로그인 이메일이 다르면 삭제 불가
+        }
+
+        // 댓글 삭제
         commentMapper.deleteComment(commentId);
+        return true;
     }
 
     @Override
